+++ config/test.exs
@@ -36,12 +36,10 @@ config :td_dd, :elasticsearch,
   es_port: 9200,
   type_name: "doc"
 
-config :td_dd, df_cache: TdPerms.MockDynamicFormCache
 config :td_dd, permission_resolver: TdDd.Permissions.MockPermissionResolver
-config :td_dd, taxonomy_cache: TdDd.MockTaxonomyCache
 config :td_dd, index_worker: TdDd.Search.MockIndexWorker
 
-config :td_perms, redis_host: "redis"
+config :td_cache, redis_host: "redis"
 
 config :td_dd, :audit_service,
   api_service: TdDdWeb.ApiServices.MockTdAuditService,