+++ lib/td_dd/search/aggregations.ex
@@ -1,8 +1,8 @@
 defmodule TdDd.Search.Aggregations do
   @moduledoc """
-    Aggregations for elasticsearch
+  Aggregations for elasticsearch
   """
-  @df_cache Application.get_env(:td_dd, :df_cache)
+  alias TdCache.TemplateCache
 
   def aggregation_terms do
     static_keywords = [
@@ -15,7 +15,7 @@ defmodule TdDd.Search.Aggregations do
     ]
 
     dynamic_keywords =
-      @df_cache.list_templates()
+      TemplateCache.list!()
       |> Enum.flat_map(&template_terms/1)
 
     (static_keywords ++ dynamic_keywords)