+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-## [Unreleased]
+## Unreleased
+
+### Changed
+
+- [TD-1618] Cache improvements. Use td-cache instead of td-perms.
+
+## [3.0.0] 2019-06-25
 
 ### Fixed
 
@@ -13,6 +19,7 @@
 
 - [TD-1793] Checks if field structure has df_content and enriches data_structure :show
 - [TD-1891] Bulk load types translation from data fields to data structures 
+- [TD-1533] Ignores search term when it is no on Aggregation
 
 ## [2.21.0] 2019-06-10